ABCModel etc. out of composer, into site
